BScreen no longer subclasses ScreenInfo, instead it contains it.
shaleh and i talked about this, and agreed that containing a screeninfo
in BScreen (which should probably be called BScreenManager for clarity) would
be better, instead of always creating 2 ScreenInfos per screen'
(one in BaseDisplay, and one in Blackbox/BScreen).

note: the code is a bit uglier now, but i feel this shows off the bad
design decisions on my part several years ago.  time for a little spring
cleaning :)