added a enter and leave notify handler for BlackboxWindow.  This event takes over control of the colormap that was done in the process_event() loop previously.  It also takes over control of auto raise.
Blackbox used to have some code borrowed from twm to implement an event
queue scanner which tried to locate LeaveNotifies in order to help avoid
useless processing.  Since it was only used for EnterNotify I simply
implemented it as a loop in the function.  Smaller, more concise code is
usually a good thing.