move class I18n into the bt namespace and adjust code as necessary.
One thing that I changed with this is the printing of the version and
usage text.  The name, version and copyright should not be localized.
the options probably shouldn't be localized either, but that is more
of a pain in the ass to change all the translations.  i'll look at
this more later.