the new Menu code.
this is a complete rewrite of the menus used in blackbox, based on the code
from moving_target (but done much cleaner).

everything works, except for tearing off menus (which will require more
research to get working properly in both the window manager and clients)

since the menu Menus are real popup menus, the Application class needs to have
extra code for delivering events in 'popup mode'  this is also complete and
working.

nothing uses the new menus yet, since they will require some changes to the
style format to work totally and correctly