as discussed with shaleh:
when focus has nowhere to go, put it on the root window (PointerRoot morexi
specifically).  this way we don't rely on the toolbar's window to be always
visible, and we can allow the toolbar to not be created if the user is
using a EWMH compliant dock (KDE/GNOME/etc.).