big API revision...
since Image is display/screen independent, Color and Texture should be as well.
when we actually need an X resource from a color/texture/image, we use
the appropriate member method, passing in the display and screen.

one noticable difference is that Color does not have a string name anymore.
instead, use Color::namedColor, which needs a display and screen to parse
the colorname into rgb values.

this ended up being a completely rewrite of Color, with has a new cache
mechanism.  Texture didn't require as many changes (thankfully).

a result of this required changing the GCCache and Pen interfaces, and
adjusting all code accordingly.

ImageControl is still alive, pending some decisions on its replacement.

bsetroot has been updated to not require imagecontrol at all.

unfortunately, this is a big update, since all code using Pen, Color, Image
and Texture had to be adjusted to the new API.