API revisions and code cleanups.
the FontCache is modelled after the ColorCache.  it will be created on
Display creation and destroyed on Display destruction.

a font no longer requires a Display object.  the object is owned by
the cache instead, so bt::Font becomes a wrapper, essentially.  this
makes the API ligher and more straight forward, and reduces the amount
of work needed to get a font loaded.