move the GCCache out of the public API.  use Pen instead.
Pen now has it's own header/implementation file, and has an API
similar to that of Color (ie. not requiring a Display argument, just a
screen).

the GCCache will be created on Display creation, and destroyed on
Display destruction.

no real changes have been made to the GCCache code. only the file name
has been changed.