added a way to get resource.double_click_interval to screen class
various touchups to the BlackboxWindow class.  Chief here is a restructure
of the button press and release events.  The if .. else tree is based on the
window the event occured on only.  This creates a more consistent UI.  xOr
submitted a similar patch a few months back and this is based largely on his
ideas.