move StackingList and BWindowGroup classes out into their own files...
add StackingList::raise() and lower(), which should be a tad bit more
efficient than simply removing and insert-/appending.

the WMLayer enum is now put together with the StackingList class,
instead of in BlackboxWindow.

A few minor cleanups are also in this diff, but there is not any
functional difference from before... the code is exactly the same
code, just moved to different places