1. use more descriptive parameter names for BlackboxWindow::setWorkspace()
2. change the way we handle GloballyActive and NoInput windows.
before we would simply refuse to focus these windows... HOWEVER, in
the case of GloballyActive, the client would expect us to send a
WM_TAKE_FOCUS message (and indicate to the user that the window is
active).  it would then move focus to some child window (we don't care
about that part though).

what does this mean?

first: things like xload can now become active... this is not in
violation of the ICCCM, since we don't explicitly set focus to the
xload window, instead we set it to the plate (the parent of the client
window).  this allows you to, for example, use bbkeys to
close/move/shade/etc. with the keyboard (since it is the active
window).

second: this should allow things like nautilus to behave properly.  I
haven't not tested this, but since nautilus is a GloballyActive
client, we need to allow users to focus it and do useful things with
it.